refactor: move modals to dedicated components
