+++ App.tsx
@@ -31,18 +31,15 @@ import SaaSClientsAdminPage from './features/SaaSClientsAdminFeature';
 import OrganizationEmailSettingsPage from './features/OrganizationEmailSettingsFeature';
 import { PageTitle, Card, Tabs, Tab, ResponsiveTable, Spinner, Button, Modal, Select as SharedSelect, Alert, Input as SharedInput, Textarea as SharedTextarea } from './components/SharedComponents';
 
+import { AddOrderCostModal } from "./components/modals/AddOrderCostModal";
 import DashboardHomePage from './features/Dashboard/Home';
 import SidebarLayout, { SidebarItem } from './components/SidebarLayout';
-import { 
+import {
     APP_NAME,
     ADMIN_APP_NAME,
-    getOrderById,
-    getOrders, // Correctly use getOrders which points to backend
-    addOrderCostItem, COST_TYPE_OPTIONS_SELECT, parseBRLCurrencyStringToNumber,
-    formatNumberToBRLCurrencyInput, formatDateBR, getClientById as getClientByIdService,
-    addClientPayment,
-} from './services/AppService';
-import { NavItem, Order, OrderCostItem, CostType, OrderStatus, ClientPayment, PaymentMethod } from './types';
+    getOrders
+} from "./services/AppService";
+import { NavItem, OrderCostItem } from "./types";
 import {
   Home,
   Users,
@@ -103,231 +100,6 @@ const ADMIN_NAV_ITEMS: NavItemWithExact[] = [
   { name: 'Auditoria & Logs', path: '/admin/audit', icon: FileText },
 ];
 
-// --- Modals (AddOrderCostModal, RegisterPaymentModal) ---
-
-// AddOrderCostModal
-interface AddOrderCostModalProps { 
-  isOpen: boolean; 
-  onClose: () => void; 
-  onSave: (costItem: OrderCostItem) => void; 
-}
-export const AddOrderCostModal: React.FC<AddOrderCostModalProps> = ({ isOpen, onClose, onSave }) => {
-    const [orderId, setOrderId] = useState('');
-    const [type, setType] = useState<CostType>(CostType.FRETE_FORNECEDOR_BLU); 
-    const [amountInput, setAmountInput] = useState('R$ 0,00');
-    const [description, setDescription] = useState('');
-    const [date, setDate] = useState(new Date().toISOString().split('T')[0]);
-    const [allOrders, setAllOrders] = useState<Order[]>([]);
-    const [error, setError] = useState<string | null>(null);
-    const [isLoadingOrders, setIsLoadingOrders] = useState(false);
-    const [isSubmitting, setIsSubmitting] = useState(false);
-
-    useEffect(() => { 
-        if (isOpen) { 
-            setIsLoadingOrders(true);
-            getOrders() // Use the corrected getOrders
-                .then(fetchedOrders => {
-                    const activeOrders = fetchedOrders.filter( o => o.status !== OrderStatus.CANCELADO && o.status !== OrderStatus.ENVIADO ).sort((a,b) => new Date(b.orderDate).getTime() - new Date(a.orderDate).getTime());
-                    setAllOrders(activeOrders);
-                    setIsLoadingOrders(false);
-                })
-                .catch(err => {
-                    console.error("Error fetching orders for cost modal:", err);
-                    setError("Falha ao carregar encomendas.");
-                    setIsLoadingOrders(false);
-                });
-            setOrderId(''); setType(CostType.FRETE_FORNECEDOR_BLU); setAmountInput('R$ 0,00'); setDescription(''); setDate(new Date().toISOString().split('T')[0]); setError(null); 
-        } 
-    }, [isOpen]);
-
-    const handleAmountChange = (e: React.ChangeEvent<HTMLInputElement>) => setAmountInput(e.target.value);
-    const handleAmountBlur = () => setAmountInput(formatNumberToBRLCurrencyInput(parseBRLCurrencyStringToNumber(amountInput)));
-    
-    const handleSubmit = async () => {
-        const amount = parseBRLCurrencyStringToNumber(amountInput);
-        if (!orderId) { setError("Selecione uma encomenda."); return; }
-        if (amount <= 0) { setError("O valor do custo deve ser maior que zero."); return; }
-        if (!description.trim() && type === CostType.OUTROS_CUSTOS) { setError("Descrição é obrigatória para 'Outros Custos'."); return; }
-        const costItemData: Omit<OrderCostItem, 'id' | 'userId'> = { orderId, type, amount, description: description.trim() || type, date, };
-        setIsSubmitting(true);
-        try {
-            const savedCostItem = await addOrderCostItem(costItemData);
-            onSave(savedCostItem);
-            onClose();
-        } catch (err) {
-            setError("Falha ao salvar custo: " + (err instanceof Error ? err.message : String(err)));
-        } finally {
-            setIsSubmitting(false);
-        }
-    };
-    const orderOptions: Array<{ value: string | number; label: string }> = [
-        { value: '', label: 'Selecione uma Encomenda Ativa...' },
-        ...allOrders.map(o => ({
-            value: o.id,
-            label: `${o.productName} ${o.model} (${o.customerName || 'Cliente não especificado'}) - Pedido: ${formatDateBR(o.orderDate)}`
-        }))
-    ];
-    return (
-        <Modal isOpen={isOpen} onClose={onClose} title="Registrar Novo Custo de Encomenda" size="lg">
-            <div className="space-y-4">
-                {error && <Alert type="error" message={error} onClose={() => setError(null)} />}
-                {isLoadingOrders ? (
-                    <Spinner />
-                ) : (
-                    <SharedSelect
-                        label="Encomenda Ativa"
-                        id="costOrderId"
-                        value={orderId}
-                        onChange={(e) => setOrderId(e.target.value)}
-                        options={orderOptions}
-                        required
-                    />
-                )}
-                <SharedSelect
-                    label="Tipo de Custo"
-                    id="costType"
-                    value={type}
-                    onChange={(e) => setType(e.target.value as CostType)}
-                    options={COST_TYPE_OPTIONS_SELECT}
-                    required
-                />
-                <SharedInput
-                    label="Valor do Custo (R$)"
-                    id="costAmount"
-                    value={amountInput}
-                    onChange={handleAmountChange}
-                    onBlur={handleAmountBlur}
-                    required
-                />
-                <SharedTextarea
-                    label={`Descrição do Custo${type === CostType.OUTROS_CUSTOS ? ' *' : ''}`}
-                    id="costDescription"
-                    value={description}
-                    onChange={(e) => setDescription(e.target.value)}
-                    rows={3}
-                    placeholder={type !== CostType.OUTROS_CUSTOS ? 'Opcional' : 'Detalhe o custo'}
-                    required={type === CostType.OUTROS_CUSTOS}
-                    aria-required={type === CostType.OUTROS_CUSTOS}
-                />
-                <SharedInput
-                    label="Data do Custo"
-                    id="costDate"
-                    type="date"
-                    value={date}
-                    onChange={(e) => setDate(e.target.value)}
-                    required
-                />
-                <div className="flex justify-end space-x-2">
-                    <Button variant="secondary" onClick={onClose} disabled={isSubmitting}>
-                        Cancelar
-                    </Button>
-                    <Button onClick={handleSubmit} isLoading={isSubmitting}>
-                        Salvar Custo
-                    </Button>
-                </div>
-            </div>
-        </Modal>
-    );
-};
-
-// RegisterPaymentModal
-interface RegisterPaymentModalProps { order: Order | null; isOpen: boolean; onClose: () => void; onPaymentSaved: (payment: ClientPayment) => void;}
-export const RegisterPaymentModal: React.FC<RegisterPaymentModalProps> = ({ order, isOpen, onClose, onPaymentSaved }) => {
-    const [paymentDate, setPaymentDate] = useState(new Date().toISOString().split('T')[0]);
-    const [amountPaidInput, setAmountPaidInput] = useState('R$ 0,00');
-    const [paymentMethodUsed, setPaymentMethodUsed] = useState<'PIX' | 'Transferência Bancária' | 'Dinheiro' | 'Cartão de Crédito' | 'Outro'>('PIX');
-    const [installments, setInstallments] = useState(2);
-    const [notes, setNotes] = useState('');
-    const [error, setError] = useState<string | null>(null);
-    const [isLoading, setIsLoading] = useState(false);
-    const [clientName, setClientName] = useState('');
-
-    useEffect(() => {
-        const fetchClientName = async () => {
-            if (order?.clientId) {
-                try {
-                    const client = await getClientByIdService(order.clientId);
-                    setClientName(client?.fullName || order.customerName);
-                } catch {
-                    setClientName(order.customerName);
-                }
-            } else if (order) {
-                setClientName(order.customerName);
-            }
-        };
-
-        if (isOpen && order) {
-            fetchClientName();
-            setPaymentDate(new Date().toISOString().split('T')[0]);
-            if (order.paymentMethod === PaymentMethod.BLU_FACILITA && order.bluFacilitaInstallments) {
-                const nextPendingInstallment = order.bluFacilitaInstallments.find( inst => inst.status === 'Pendente' || inst.status === 'Atrasado' || inst.status === 'Pago Parcialmente' );
-                const suggestedAmount = nextPendingInstallment ? (nextPendingInstallment.amount - (nextPendingInstallment.amountPaid || 0)) : (order.installmentValue || 0);
-                setAmountPaidInput(formatNumberToBRLCurrencyInput(suggestedAmount > 0 ? suggestedAmount : 0));
-            } else { setAmountPaidInput(formatNumberToBRLCurrencyInput(order.sellingPrice || 0)); }
-            setPaymentMethodUsed('PIX'); setInstallments(2); setNotes(''); setError(null);
-        }
-    }, [isOpen, order]);
-
-    const handleAmountChange = (e: React.ChangeEvent<HTMLInputElement>) => setAmountPaidInput(e.target.value);
-    const handleAmountBlur = () => setAmountPaidInput(formatNumberToBRLCurrencyInput(parseBRLCurrencyStringToNumber(amountPaidInput)));
-    const handleSubmit = async () => {
-        if (!order) { setError("Encomenda não especificada."); return; }
-        const amountPaid = parseBRLCurrencyStringToNumber(amountPaidInput);
-        if (amountPaid <= 0) { setError("O valor pago deve ser maior que zero."); return; }
-        setError(null);
-        setIsLoading(true);
-        try {
-            const paymentData: Omit<ClientPayment, 'id'|'userId'> = {
-                orderId: order.id,
-                paymentDate,
-                amountPaid,
-                paymentMethodUsed,
-                installments: paymentMethodUsed === 'Cartão de Crédito' ? installments : undefined,
-                notes
-            };
-            const savedPayment = await addClientPayment(paymentData);
-            onPaymentSaved(savedPayment);
-            onClose();
-        } catch (err) {
-            setError("Falha ao registrar pagamento: " + (err instanceof Error ? err.message : String(err)));
-        } finally {
-            setIsLoading(false);
-        }
-    };
-    const paymentMethodOptions = [
-        { value: 'PIX', label: 'PIX' },
-        { value: 'Transferência Bancária', label: 'Transferência Bancária' },
-        { value: 'Dinheiro', label: 'Dinheiro' },
-        { value: 'Cartão de Crédito', label: 'Cartão de Crédito' },
-        { value: 'Outro', label: 'Outro' },
-    ];
-    return (
-        <Modal isOpen={isOpen} onClose={onClose} title={`Registrar Recebimento`} size="lg">
-            <div className="space-y-4">
-                {error && <Alert type="error" message={error} onClose={() => setError(null)} />}
-                <p><strong>Cliente:</strong> {clientName}</p>
-                <p><strong>Produto:</strong> {order?.productName} {order?.model}</p>
-                <SharedInput label="Data do Recebimento" id="paymentDate" type="date" value={paymentDate} onChange={(e) => setPaymentDate(e.target.value)} required />
-                <SharedInput label="Valor Recebido (R$)" id="amountPaid" value={amountPaidInput} onChange={handleAmountChange} onBlur={handleAmountBlur} required />
-                <SharedSelect label="Forma de Pagamento Usada" id="paymentMethodUsed" value={paymentMethodUsed} onChange={(e) => setPaymentMethodUsed(e.target.value as any)} options={paymentMethodOptions} required />
-                {paymentMethodUsed === 'Cartão de Crédito' && (
-                    <SharedSelect
-                        label="Número de Parcelas"
-                        id="installments"
-                        value={String(installments)}
-                        onChange={(e) => setInstallments(parseInt(e.target.value))}
-                        options={Array.from({ length: 11 }, (_, i) => ({ value: i + 2, label: `${i + 2}x` }))}
-                    />
-                )}
-                <SharedTextarea label="Observações (Opcional)" id="paymentNotes" value={notes} onChange={(e) => setNotes(e.target.value)} rows={2} />
-                <div className="flex justify-end space-x-2">
-                    <Button variant="secondary" onClick={onClose} disabled={isLoading}>Cancelar</Button>
-                    <Button onClick={handleSubmit} isLoading={isLoading}>Salvar Recebimento</Button>
-                </div>
-            </div>
-        </Modal>
-    );
-};
 
 
 const Sidebar: React.FC<{isOpen: boolean; setIsOpen: (isOpen: boolean) => void;}> = ({ isOpen, setIsOpen }) => {
